requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
django-bootstrap4==22.1
django-recaptcha==3.0.0
idna==3.3
requests==2.27.1
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8